filter_jwplayer: Use CDN version for cloud-hosted mode.
As there is no way at the moment to use the cloud player in AMD module, and
the concept has changed in a way that JW Player supplies pre-configured
player for the given token, it has been decided to switch to CDN version of
JW Player and use it as "cloud-hosting" mode. This makes license key
required for both cloud and self-hosted, as cloud in this case is similar
to self-hosted, but supplied using JW Player CDN. Also, token supplying is
becoming redundant in this case, and apprarently makes process less
confusing, because now to get the token user is required to precofigure the
player using JW Player dashboard.

There is a potential to support cloud player in the future, but it should
be something that is using JW Platform API to fetch the player which is
entirely pre-configured using JW Player dashboard (see JW Player Wordpress
plugin), this does not fit this plugin really which has reach player
customisation features, perhaps need a separate one with mininum settings.